--- cupon-app/src/main/java/com/meli/cupon/domain/excepcion/DomainException.java
@@ -1,8 +1,0 @@
-package com.meli.cupon.domain.excepcion;
-
-public abstract class DomainException extends RuntimeException {
-
-    public DomainException(String message) {
-        super(message);
-    }
-}
