--- cupon-app/src/main/java/com/meli/cupon/domain/model/entities/IdItem.java
@@ -1,10 +1,0 @@
-package com.meli.cupon.domain.model.entities;
-
-import org.springframework.util.Assert;
-
-public record IdItem(String valor) {
-
-    public IdItem {
-        Assert.hasText(valor, "Invalid value.");
-    }
-}
