--- cupon-app/src/main/java/com/meli/cupon/domain/model/entities/Item.java
@@ -1,10 +1,0 @@
-package com.meli.cupon.domain.model.entities;
-
-import java.math.BigDecimal;
-
-public record Item(IdItem itemId, BigDecimal value) {
-
-    public Item {
-        if (value.signum() <= 0) throw new IllegalArgumentException("Invalid value.");
-    }
-}
