--- cupon-app/src/main/java/com/meli/cupon/domain/services/BaseService.java
@@ -1,5 +1,0 @@
-package com.meli.cupon.domain.services;
-
-public class BaseService {
-
-}
