--- cupon-app/src/main/java/com/meli/cupon/infrastruture/config/ApplicationConfig.java
@@ -1,8 +1,0 @@
-package com.meli.cupon.infrastruture.config;
-
-import org.springframework.context.annotation.Configuration;
-
-@Configuration
-public class ApplicationConfig {
-
-}
