--- cupon-app/src/main/java/com/meli/cupon/infrastruture/remote/items/ItemDto.java
@@ -1,8 +1,0 @@
-package com.meli.cupon.infrastruture.remote.items;
-
-import java.math.BigDecimal;
-
-public class ItemDto {
-
-    public BigDecimal price;
-}
