+++ cupon-app/src/main/java/com/meli/login/domain/excepcion/DomainException.java
@@ -1,0 +1,8 @@
+package com.meli.login.domain.excepcion;
+
+public abstract class DomainException extends RuntimeException {
+
+    public DomainException(String message) {
+        super(message);
+    }
+}
