+++ cupon-app/src/main/java/com/meli/login/domain/model/entities/IdItem.java
@@ -1,0 +1,10 @@
+package com.meli.login.domain.model.entities;
+
+import org.springframework.util.Assert;
+
+public record IdItem(String valor) {
+
+    public IdItem {
+        Assert.hasText(valor, "Invalid value.");
+    }
+}
