+++ cupon-app/src/main/java/com/meli/login/domain/services/BaseService.java
@@ -1,0 +1,5 @@
+package com.meli.login.domain.services;
+
+public class BaseService {
+
+}
