+++ cupon-app/src/main/java/com/meli/login/infrastruture/config/ApplicationConfig.java
@@ -1,0 +1,8 @@
+package com.meli.login.infrastruture.config;
+
+import org.springframework.context.annotation.Configuration;
+
+@Configuration
+public class ApplicationConfig {
+
+}
