+++ cupon-app/src/main/java/com/meli/login/infrastruture/controler/IndexController.java
@@ -1,0 +1,15 @@
+package com.meli.login.infrastruture.controler;
+
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RestController;
+import reactor.core.publisher.Mono;
+
+@RestController
+public class IndexController {
+
+    @GetMapping("/up")
+    public Mono<ResponseEntity<String>> up() {
+        return Mono.just(ResponseEntity.ok("The cupon-app is up and running"));
+    }
+}
