+++ cupon-app/src/main/java/com/meli/login/infrastruture/remote/items/ItemDto.java
@@ -1,0 +1,8 @@
+package com.meli.login.infrastruture.remote.items;
+
+import java.math.BigDecimal;
+
+public class ItemDto {
+
+    public BigDecimal price;
+}
