--- cupon-app/src/test/groovy/com/meli/cupon/domain/services/CouponServiceSpec.groovy
@@ -1,9 +1,0 @@
-package com.meli.cupon.domain.services
-
-import spock.lang.Ignore
-import spock.lang.Specification
-
-@Ignore
-class CouponServiceSpec extends Specification {
-
-}
