+++ cupon-app/src/test/groovy/com/meli/login/domain/services/CouponServiceSpec.groovy
@@ -1,0 +1,9 @@
+package com.meli.login.domain.services
+
+import spock.lang.Ignore
+import spock.lang.Specification
+
+@Ignore
+class CouponServiceSpec extends Specification {
+
+}
